// packages/client/src/config/constants.tsx
export enum Paths {
  App = '/',
  SignIn = '/sign-in',
  SignUp = '/sign-up',
  Profile = '/profile',
  Main = '/main',
  Game = '/game',
  Leaderboard = '/leaderboard',
  Forum = '/forum',
  Topic = '/topic/:forumId',
  Error404 = '/404',
  Error500 = '/500',
}

export const LOCAL_CLIENT_PORT = 3000

export const LOCAL_SERVER_PORT = 3001

export const YANDEX_API_HOST = 'https://ya-praktikum.tech/api/v2'